** ADDING NEW PNG FILES **
1) I am thinking of changing the user and cpu gambits into actual functions. (ref: https://youtu.be/EPwszp6Ecgs)

In the GUI, I have assigned lambda functions to the three buttons. However, I am unsure as to where and how to place the gambit functions--in the main, logic file, or GUI?--and how to link the files together without creating unnecessary loops/imports.

Where to put these???
    def cpu_choice():
        choice = random.choice(gambits)
        return choice

    def user_choice(choice):
        global click

        cpu_current = cpu_choice()

2) Added quit button to GUI

3) Added score variable to main

4) Commented out code for non-gui game

5) Added all sprites--> Next task: Put all sprites into a single list, for the GUI's label changes.

My "resize_sprites" function DOES NOT WORK in the way I hoped. Instead, I will create an indexed list (or not?) and put in the "file" slot for card_background_photo --> add the conditions to the if statements in the game_logic file? or the main?